requirements/base.txt
pytz==2021.3  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==9.0.1  # https://github.com/python-pillow/Pillow
rcssmin==1.1.0  # https://github.com/ndparker/rcssmin
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
redis==4.1.4  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.3  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.2.1  # https://github.com/celery/django-celery-beat
flower==1.0.0  # https://github.com/mher/flower
django-autoslug==1.9.8  # https://github.com/justinmayer/django-autoslug/

# Django
# ------------------------------------------------------------------------------
django==3.2.11  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.47.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.13.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==3.1  # https://github.com/django-compressor/django-compressor
django-redis==5.2.0  # https://github.com/jazzband/django-redis
crispy-tailwind==0.5.0 # https://www.accordbox.com/blog/render-django-form-with-tailwind-css-style/
# djangosaml2idp==0.7.2 # https://djangosaml2idp.readthedocs.io/en/latest/index.html
djangosaml2idp@git+https://github.com/Musab0/djangosaml2idp.git@db208dc
django-iprestrict-redux==1.9.0 #https://django-iprestrict-redux.readthedocs.io/en/latest/
djangorestframework
djangorestframework-simplejwt
django-cors-headers

# Project
# ------------------------------------------------------------------------------
PyGithub==1.55   # https://github.com/PyGithub/PyGithub
slackclient==2.9.3  # https://pypi.org/project/slackclient
django-ckeditor==6.2.0 # https://pypi.org/project/django-ckeditor/
pandas==1.3.5 # https://pypi.org/project/pandas/
